# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: BizCore Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /items/{itemId}/recipe:
    get:
      operationId: getRecipe
      tags: [recipes]
      summary: Get the recipe for a menu item
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Recipe with ingredients
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecipeDetail"
        "404":
          description: No recipe found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: upsertRecipe
      tags: [recipes]
      summary: Create or replace the recipe for a menu item
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertRecipeBody"
      responses:
        "200":
          description: Recipe saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecipeDetail"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    RecipeItem:
      type: object
      properties:
        id:
          type: integer
        recipeId:
          type: integer
        ingredientVariantId:
          type: integer
        quantity:
          type: string
        unit:
          type: ["string", "null"]
        ingredientName:
          type: ["string", "null"]
        variantName:
          type: ["string", "null"]
      required:
        - id
        - recipeId
        - ingredientVariantId
        - quantity
    RecipeDetail:
      type: object
      properties:
        id:
          type: integer
        menuItemId:
          type: integer
        name:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/RecipeItem"
      required:
        - id
        - menuItemId
        - items
    UpsertRecipeBody:
      type: object
      properties:
        name:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        items:
          type: array
          items:
            type: object
            properties:
              ingredientVariantId:
                type: integer
              quantity:
                type: string
              unit:
                type: ["string", "null"]
            required:
              - ingredientVariantId
              - quantity
      required:
        - items